% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_management_step.R
\name{poststep1}
\alias{poststep1}
\title{poststep1 - Control chart for post management (I-MR)}
\usage{
poststep1(x)
}
\arguments{
\item{x}{: i-mr control chart data (individual data)}
}
\value{
: i-mr control chart
}
\description{
poststep1 - Control chart for post management (I-MR)
}
\examples{
 poststep1(pss4$throw_Distance)
}
