% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_Check.R
\name{pre_check}
\alias{pre_check}
\title{pre_check - install required packages}
\usage{
pre_check()
}
\value{
Install and load required packages-'qcc','SixSigma','randomForest','devtools' package
}
\description{
pre_check - install required packages
}
\examples{
pre_check(  )
}
